[enhancement] add behavior predictor (#3)
🔮 Predicting the Future, One Commit at a Time!

Hey Team,

I've just pushed the future into our repo, quite literally! Introducing
our brand-new Behavior Predictor - because why wait for users to
surprise us when we can surprise them with our clairvoyance?

This shiny new feature is like a crystal ball but with more algorithms
and fewer mystical vibes. It's designed to guess what our users will do
next, hopefully with more accuracy than my attempts to predict lottery
numbers.

Before you ask – no, it can't predict when we'll go back to the office
or what's for lunch tomorrow, but it sure will add some wizardry to our
user experience!

So, let's merge this and stay one step ahead, or at least pretend to be!

Cheers,
Christian

P.S. If this thing predicts the end of the world, let's agree to keep it
hush-hush. We're developers, not meteorologists! 🌎💫